feat: adiciona componente Relogio
